+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 #### V0.1.7
 - 增加 alz.moving_prediction 方法
+- ml.WaveNet 现在返回和输入序列长度相同的特征
+- ml.WaveNet 增加 valid_loss 参数，开启后只在有效观测区间计算loss
 
 #### V0.1.6
 - 增加 io.DarkSkyAPI